+++ lib/kernel/kernel.h
@@ -5,7 +5,7 @@
 void keyboard_handler(void);
 
 // load the ISR
-void load_idt(uint32_t *idt_ptr);
+void load_idt(uintptr_t *idt_ptr);
 
 // checks to make sure we can run on this hardware
 bool cpuid_supported(void);
